Make a separate `translate_bodies` module
